+++ src/components/Menu/Menu.tsx
@@ -1,16 +1,17 @@
-import React, { useReducer, useState } from "react";
-import * as S from "./style";
-
+import React, { useState } from "react";
 import { BiMenu } from "react-icons/bi";
-import { settingReducer, initialSetting } from "./settingReducer";
+
+import * as S from "./style";
 import { useToggleBtns } from "./menuFunc/useToggleBtns";
-import TitleInput from "./TitleInput";
-import ToggleTitle from "./ToggleTitle";
-import ToggleSearch from "./ToggleSearch";
-import ToggleMultiSelect from "./ToggleMultiSelect";
-import ToggleSelectedItem from "./ToggleSelectedItem";
-import SetItemSize from "./SetItemSize";
-import SetBoxSize from "./SetBoxSize";
+import {
+  TitleInput,
+  ToggleTitle,
+  ToggleSearch,
+  ToggleMultiSelect,
+  ToggleSelectedItem,
+  SetItemSize,
+  SetBoxSize,
+} from "./MenuComponents";
 import { Props } from "./types";
 
 function Menu({ settingState, dispatch }: Props) {
