+++ src/components/Menu/MenuComponents/SetItemSize/types.ts
@@ -1,0 +1,7 @@
+import { ClickUlEventFunc } from "../../menuFunc/types";
+import { Setting } from "../../settingReducer/types";
+
+export interface Props {
+  settingState: Setting;
+  handleUlClick: ClickUlEventFunc;
+}
