+++ src/components/Menu/MenuComponents/TitleInput/types.ts
@@ -1,5 +1,5 @@
-import { InputChangeFunc } from "../menuFunc/types";
-import { Setting } from "../settingReducer/types";
+import { InputChangeFunc } from "../../menuFunc/types";
+import { Setting } from "../../settingReducer/types";
 
 export interface Props {
   settingState: Setting;
