+++ src/components/Menu/MenuComponents/ToggleMultiSelect/types.ts
@@ -1,0 +1,7 @@
+import { ToggleFunc } from "../../menuFunc/types";
+import { Setting } from "../../settingReducer/types";
+
+export interface Props {
+  settingState: Setting;
+  handleToggleMultiSelect: ToggleFunc;
+}
