+++ src/components/Menu/MenuComponents/ToggleTitle/ToggleTitle.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Props } from "./types";
-import { ToggleDiv } from "../style";
+import { ToggleDiv } from "../../style";
 
 const ToggleTitle = ({ settingState, handleToggleShowTitle }: Props) => {
   const { showTitle } = settingState;
