+++ src/components/Menu/MenuComponents/index.tsx
@@ -1,0 +1,7 @@
+export { default as SetBoxSize } from "./SetBoxSize";
+export { default as SetItemSize } from "./SetItemSize";
+export { default as TitleInput } from "./TitleInput";
+export { default as ToggleMultiSelect } from "./ToggleMultiSelect";
+export { default as ToggleSearch } from "./ToggleSearch";
+export { default as ToggleSelectedItem } from "./ToggleSelectedItem";
+export { default as ToggleTitle } from "./ToggleTitle";
