--- src/components/Menu/SetItemSize/types.ts
@@ -1,7 +1,0 @@
-import { Setting } from "../settingReducer/types";
-import { ClickUlEventFunc } from "../menuFunc/types";
-
-export interface Props {
-  settingState: Setting;
-  handleUlClick: ClickUlEventFunc;
-}
