--- src/components/Menu/ToggleSelectedItem/types.ts
@@ -1,7 +1,0 @@
-import { ToggleFunc } from "../menuFunc/types";
-import { Setting } from "../settingReducer/types";
-
-export interface Props {
-  settingState: Setting;
-  handleToggleShowSelectedItem: ToggleFunc;
-}
